+++ src/main/java/de/mossgrabers/convertwithmoss/core/ConverterBackend.java
@@ -19,7 +19,6 @@ import de.mossgrabers.convertwithmoss.core.model.IEnvelope;
 import de.mossgrabers.convertwithmoss.core.model.IGroup;
 import de.mossgrabers.convertwithmoss.core.model.ISampleZone;
 import de.mossgrabers.convertwithmoss.core.model.implementation.DefaultEnvelope;
-import de.mossgrabers.convertwithmoss.file.CSVRenameFile;
 import de.mossgrabers.convertwithmoss.format.ableton.AbletonCreator;
 import de.mossgrabers.convertwithmoss.format.ableton.AbletonDetector;
 import de.mossgrabers.convertwithmoss.format.akai.akp.AkpDetector;
@@ -82,12 +81,8 @@ public class ConverterBackend
 
     private IDetector<?>                   detector;
     private ICreator<?>                    creator;
-    private File                           outputFolder;
-    private CSVRenameFile                  csvRenameFile;
-    private String                         libraryName;
-    private boolean                        wantsMultipleFiles;
+    private DetectSettings                 detectionSettings;
     private boolean                        onlyAnalyse;
-    private boolean                        createFolderStructure;
 
     private final List<IMultisampleSource> collectedPresetSources      = new ArrayList<> ();
     private final List<IPerformanceSource> collectedPerformanceSources = new ArrayList<> ();
@@ -188,26 +183,16 @@ public class ConverterBackend
      *
      * @param detector The file detector
      * @param creator The file creator
-     * @param sourceFolder The folder where to start the detection process
-     * @param outputFolder Where to write the result to
-     * @param csvRenameFile If renaming is required
-     * @param libraryName The name to use in case that a library will be created
+     * @param detectionSettings The settings for the detection process
      * @param detectPerformances If true, performances are detected otherwise presets
-     * @param wantsMultipleFiles True, if all files should be returned at once
-     * @param createFolderStructure True, if the source folder structure should be replicated in the
-     *            output folder
      * @param onlyAnalyse True, if no files should be created
      */
-    public void detect (final IDetector<?> detector, final ICreator<?> creator, final File sourceFolder, final File outputFolder, final CSVRenameFile csvRenameFile, final String libraryName, final boolean detectPerformances, final boolean wantsMultipleFiles, final boolean createFolderStructure, final boolean onlyAnalyse)
+    public void detect (final IDetector<?> detector, final ICreator<?> creator, final DetectSettings detectionSettings, final boolean detectPerformances, final boolean onlyAnalyse)
     {
         this.detector = detector;
         this.creator = creator;
-        this.outputFolder = outputFolder;
-        this.csvRenameFile = csvRenameFile;
-        this.libraryName = libraryName;
-        this.wantsMultipleFiles = wantsMultipleFiles;
+        this.detectionSettings = detectionSettings;
         this.onlyAnalyse = onlyAnalyse;
-        this.createFolderStructure = createFolderStructure;
 
         this.collectedPresetSources.clear ();
         this.collectedPerformanceSources.clear ();
@@ -215,7 +200,7 @@ public class ConverterBackend
         this.notifier.log ("TITLE");
         this.notifier.log ("IDS_NOTIFY_DETECTING", detector.getName (), creator.getName ());
         this.creator.clearCancelled ();
-        this.detector.detect (sourceFolder, new MultisampleSourceConsumer (), new PerformanceSourceConsumer (), detectPerformances);
+        this.detector.detect (detectionSettings.sourceFolder, new MultisampleSourceConsumer (), new PerformanceSourceConsumer (), detectPerformances);
     }
 
 
@@ -241,13 +226,13 @@ public class ConverterBackend
             {
                 if (!this.collectedPresetSources.isEmpty ())
                 {
-                    final String name = this.getPresetLibraryName (this.collectedPresetSources, this.libraryName);
-                    this.creator.createPresetLibrary (this.outputFolder, this.collectedPresetSources, name);
+                    final String name = this.getPresetLibraryName (this.collectedPresetSources, this.detectionSettings.libraryName);
+                    this.creator.createPresetLibrary (this.detectionSettings.outputFolder, this.collectedPresetSources, name);
                 }
                 else if (!this.collectedPerformanceSources.isEmpty ())
                 {
-                    final String name = this.getPerformanceLibraryName (this.collectedPerformanceSources, this.libraryName);
-                    this.creator.createPerformanceLibrary (this.outputFolder, this.collectedPerformanceSources, name);
+                    final String name = this.getPerformanceLibraryName (this.collectedPerformanceSources, this.detectionSettings.libraryName);
+                    this.creator.createPerformanceLibrary (this.detectionSettings.outputFolder, this.collectedPerformanceSources, name);
                 }
             }
             catch (final IOException | RuntimeException | OutOfMemoryError ex)
@@ -264,11 +249,9 @@ public class ConverterBackend
         if (this.detector.isCancelled ())
             return;
 
-        ensureSafeSampleFileNames (multisampleSource);
-        this.applyRenaming (multisampleSource);
-        this.applyDefaultEnvelope (multisampleSource);
+        this.processSource (multisampleSource);
 
-        if (this.wantsMultipleFiles)
+        if (this.detectionSettings.wantsMultipleFiles)
         {
             if (!this.onlyAnalyse)
                 this.collectedPresetSources.add (multisampleSource);
@@ -286,7 +269,7 @@ public class ConverterBackend
 
         try
         {
-            final File multisampleOutputFolder = calcOutputFolder (this.outputFolder, multisampleSource.getSubPath (), this.createFolderStructure);
+            final File multisampleOutputFolder = calcOutputFolder (this.detectionSettings.outputFolder, multisampleSource.getSubPath (), this.detectionSettings.createFolderStructure);
             this.creator.createPreset (multisampleOutputFolder, multisampleSource);
         }
         catch (final NoSuchFileException | FileNotFoundException ex)
@@ -309,16 +292,10 @@ public class ConverterBackend
         if (instrumentSources.isEmpty ())
             return;
 
-        IMultisampleSource multisampleSource;
         for (final IInstrumentSource instrumentSource: instrumentSources)
-        {
-            multisampleSource = instrumentSource.getMultisampleSource ();
-            ensureSafeSampleFileNames (multisampleSource);
-            this.applyRenaming (multisampleSource);
-            this.applyDefaultEnvelope (multisampleSource);
-        }
+            this.processSource (instrumentSource.getMultisampleSource ());
 
-        if (this.wantsMultipleFiles)
+        if (this.detectionSettings.wantsMultipleFiles)
         {
             if (!this.onlyAnalyse)
                 this.collectedPerformanceSources.add (performanceSource);
@@ -334,7 +311,7 @@ public class ConverterBackend
 
         try
         {
-            final File multisampleOutputFolder = calcOutputFolder (this.outputFolder, instrumentSources.get (0).getMultisampleSource ().getSubPath (), this.createFolderStructure);
+            final File multisampleOutputFolder = calcOutputFolder (this.detectionSettings.outputFolder, instrumentSources.get (0).getMultisampleSource ().getSubPath (), this.detectionSettings.createFolderStructure);
             this.creator.createPerformance (multisampleOutputFolder, performanceSource);
         }
         catch (final NoSuchFileException | FileNotFoundException ex)
@@ -345,6 +322,25 @@ public class ConverterBackend
         {
             this.notifier.logError ("IDS_NOTIFY_SAVE_FAILED", ex);
         }
+    }
+
+
+    private void processSource (final IMultisampleSource multisampleSource)
+    {
+        ensureSafeSampleFileNames (multisampleSource);
+        this.processSamples (multisampleSource);
+        this.applyRenaming (multisampleSource);
+        this.applyDefaultEnvelope (multisampleSource);
+    }
+
+
+    private void processSamples (final IMultisampleSource multisampleSource)
+    {
+        if (this.onlyAnalyse)
+            return;
+
+        // TODO Auto-generated method stub
+
     }
 
 
@@ -393,11 +389,11 @@ public class ConverterBackend
      */
     private void applyRenaming (final IMultisampleSource multisampleSource)
     {
-        if (this.csvRenameFile == null || this.csvRenameFile.isEmpty ())
+        if (this.detectionSettings.csvRenameFile == null || this.detectionSettings.csvRenameFile.isEmpty ())
             return;
 
         final String sourceName = multisampleSource.getName ();
-        final String targetName = this.csvRenameFile.getMapping (sourceName);
+        final String targetName = this.detectionSettings.csvRenameFile.getMapping (sourceName);
         if (targetName != null)
         {
             this.notifier.log ("IDS_NOTIFY_RENAMING_SOURCE_TO", sourceName, targetName);
@@ -435,14 +431,14 @@ public class ConverterBackend
 
     private String getPresetLibraryName (final List<IMultisampleSource> multisampleSources, final String libraryName)
     {
-        final String name = this.wantsMultipleFiles && !libraryName.isEmpty () ? libraryName : multisampleSources.get (0).getName ();
+        final String name = this.detectionSettings.wantsMultipleFiles && !libraryName.isEmpty () ? libraryName : multisampleSources.get (0).getName ();
         return AbstractCreator.createSafeFilename (name);
     }
 
 
     private String getPerformanceLibraryName (final List<IPerformanceSource> performanceSources, final String libraryName)
     {
-        final String name = this.wantsMultipleFiles && !libraryName.isEmpty () ? libraryName : performanceSources.get (0).getName ();
+        final String name = this.detectionSettings.wantsMultipleFiles && !libraryName.isEmpty () ? libraryName : performanceSources.get (0).getName ();
         return AbstractCreator.createSafeFilename (name);
     }