+++ src/main/java/de/mossgrabers/convertwithmoss/ui/ProcessingDialog.java
@@ -11,6 +11,7 @@ import java.util.List;
 import de.mossgrabers.tools.ui.AbstractDialog;
 import de.mossgrabers.tools.ui.ControlFunctions;
 import de.mossgrabers.tools.ui.TraversalManager;
+import de.mossgrabers.tools.ui.panel.BasePanel;
 import de.mossgrabers.tools.ui.panel.BoxPanel;
 import javafx.geometry.Orientation;
 import javafx.scene.control.CheckBox;
@@ -45,10 +46,8 @@ public class ProcessingDialog extends AbstractDialog
     public CheckBox                normalizeCheckbox;
     /** Check-box for enabling making all samples mono. */
     public CheckBox                makeMonoCheckbox;
-    /** Check-box for enabling the truncate-sample-before-playback-start option. */
-    public CheckBox                truncateSampleBeforePlaybackStartCheckbox;
-    /** Check-box for enabling the truncate-sample-after-loop-end option. */
-    public CheckBox                truncateSampleAfterLoopEndCheckbox;
+    /** Check-box for enabling the trim start/end option. */
+    public CheckBox                trimSample;
     /** Text field for the maximum number of samples. */
     public TextField               maxSamplesField;
     /** Combo-box for the target bit-depth. */
@@ -85,22 +84,19 @@ public class ProcessingDialog extends AbstractDialog
 
         panel.createSeparator ("@IDS_PROCESSING_MINIMIZE_HEADER");
         this.makeMonoCheckbox = panel.createCheckBox ("@IDS_PROCESSING_MONO", "@IDS_PROCESSING_MONO_TOOLTIP");
-        this.truncateSampleBeforePlaybackStartCheckbox = panel.createCheckBox ("@IDS_PROCESSING_TRUNCATE_START", "@IDS_PROCESSING_TRUNCATE_START_TOOLTIP");
-        this.truncateSampleAfterLoopEndCheckbox = panel.createCheckBox ("@IDS_PROCESSING_TRUNCATE_AFTER_LOOP", "@IDS_PROCESSING_TRUNCATE_AFTER_LOOP_TOOLTIP");
+        this.trimSample = panel.createCheckBox ("@IDS_PROCESSING_TRUNCATE_START", "@IDS_PROCESSING_TRUNCATE_START_TOOLTIP");
         this.maxSamplesField = panel.createPositiveIntegerField ("@IDS_PROCESSING_MAX_SAMPLES", "@IDS_PROCESSING_MAX_SAMPLES_TOOLTIP");
+        BasePanel.limitToNumbers (this.maxSamplesField);
         this.reduceBitDepthCombobox = panel.createComboBox ("@IDS_PROCESSING_REDUCE_BIT_DEPTH", "@IDS_PROCESSING_REDUCE_BIT_DEPTH_TOOLTIP", BIT_DEPTH);
         this.reduceFrequencyCombobox = panel.createComboBox ("@IDS_PROCESSING_REDUCE_FREQUENCY", "@IDS_PROCESSING_REDUCE_FREQUENCY_TOOLTIP", FREQ_RESOLUTiON);
-
         this.setButtons ("@IDS_SETTINGS_DLG_OK", "@IDS_SETTINGS_DLG_CANCEL");
 
         this.traversalManager.add (this.normalizeCheckbox);
         this.traversalManager.add (this.makeMonoCheckbox);
-        this.traversalManager.add (this.truncateSampleBeforePlaybackStartCheckbox);
-        this.traversalManager.add (this.truncateSampleAfterLoopEndCheckbox);
+        this.traversalManager.add (this.trimSample);
         this.traversalManager.add (this.maxSamplesField);
         this.traversalManager.add (this.reduceBitDepthCombobox);
         this.traversalManager.add (this.reduceFrequencyCombobox);
-
         this.traversalManager.add (this.getOKButton ());
         this.traversalManager.add (this.getCancelButton ());